Address broken package building due to tempdirs
Recently, support was added to build packages in temporary directories.
This inadvertently led to a bug wherein when the build is finished, and
the tempdir is removed, the packages were removed with it.

This fix adds the output_dir setting.  If set, the packages will be
output to the directory specified, else to the cwd of the giftwrap
process.