Update Vagrantfile to make use of OpenStack provider
Updating the Vagrantfile so that builds may be run in an OpenStack
cleanroom instance as well as locally via virtualbox. In order to
properly configure the vagrant instance, a number of environment
variables are utilized. First, all normal OpenStack auth envronment
vars are used to create the instance, and then a host of
GIFTWRAP_OS_* vars are added to control things like the image being
spun up, the networks it is attached to, etc.

A postbuild mechanism is added to allow one to do things like ship
artifacts to a repo, Swift, etc.