Introduce upper constraints
This is a global setting that will apply across every project, a
constraint on the versions of things that can be installed via pip.

This is a first pass, adding basic functionality, by allowing a single
constraint URL to be provided. Future iterations of this feature may be
tied to a "superrepo" that contains all the projects _and_ the
constraints for those projects all in one place.

Change-Id: Ib5c2dd441b9294b775e755e595486a671d8de57f