Make buildspec take superrepo
Setting a superrepo in the build spec will cause projects to be
populated using the superrepo contents. Currently this also means we get
a double-clone if we use this to build packages, but we can optimize
this later by recognizing we already have what we want and skipping the
clone.

Change-Id: Ib8bec440768a2447e576dfffbd880130f6db1fdb