Fix test decorators
These tests aren't actually run by nose because the private decorator
doesn't properly wrap its test. This is also a useful pattern, and so we
move it into its own file (note that even better would be to use
testtools and fixtures, but that's a bigger refactor).

In addition, we don't need to use the git CLI, when we have gitpython,
which produces cleaner code and doesn't spray stuff all over our stderr.

Finally, the outdir creating decorator wasn't as reusable as I thought
it would be, so removing it for inline code for now.

Change-Id: Ibe1b993235b68a1a59bc398ded9c822a02475570