Edit constraints to remove project
Pip constraints have a limitation that you cannot constrain the project
you are actually installing. So the constraints file must be editted
just to remove the project if it happens to be listed.

We refactor the builders here so we can actually pass in the project
name, and also to move duplicate code into a shared parent class.

Change-Id: I83337ee08f2405fc0c648da4a142c654e68c0b98